make RefPath and RefResponse extend Path and Response respectively
