+++ modules/swagger-core/src/main/java/io/swagger/util/Yaml.java
@@ -1,38 +1,15 @@
 package io.swagger.util;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.util.DefaultPrettyPrinter;
-import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.databind.module.SimpleModule;
-import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
-import io.swagger.models.Model;
-import io.swagger.models.Path;
-import io.swagger.models.Response;
-import io.swagger.models.auth.SecuritySchemeDefinition;
-import io.swagger.models.parameters.Parameter;
-import io.swagger.models.properties.Property;
 
 public class Yaml {
     static ObjectMapper mapper;
 
     public static ObjectMapper mapper() {
-        if (mapper == null) {
-            mapper = new ObjectMapper(new YAMLFactory());
-            // mapper.registerModule(new JodaModule());
-            SimpleModule module = new SimpleModule();
-            module.addDeserializer(Path.class, new PathDeserializer());
-            module.addDeserializer(Response.class, new ResponseDeserializer());
-            module.addDeserializer(Model.class, new ModelDeserializer());
-            module.addDeserializer(Property.class, new PropertyDeserializer());
-            module.addDeserializer(Parameter.class, new ParameterDeserializer());
-            module.addDeserializer(SecuritySchemeDefinition.class, new SecurityDefinitionDeserializer());
-            mapper.registerModule(module);
-            mapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
-            mapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false);
-            mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+        if(mapper == null) {
+            mapper = ObjectMapperFactory.createYaml();
         }
         return mapper;
     }
