+++ modules/swagger-core/src/test/scala/ResponseExamplesTest.scala
@@ -1,4 +1,4 @@
-import io.swagger.models.{ResponseImpl, Response}
+import io.swagger.models.{Response}
 import io.swagger.util.Json
 import org.junit.runner.RunWith
 import org.scalatest.{FlatSpec, Matchers}
@@ -7,7 +7,7 @@ import org.scalatest.junit.JUnitRunner
 @RunWith(classOf[JUnitRunner])
 class ResponseExamplesTest extends FlatSpec with Matchers {
   it should "create a response" in {
-    val response = new ResponseImpl()
+    val response = new Response()
       .example("application/json", """{"name":"Fred","id":123456"}""")
     Json.mapper().writeValueAsString(response) should equal( """{"examples":{"application/json":"{\"name\":\"Fred\",\"id\":123456\"}"}}""")
   }
